feat: refactor external comments to React Server Components with package distribution support
## What Changed
- Created new React Server Component version (ExternalCommentsServer) for better performance
- Extracted shared data fetching utilities to lib/external-comments.ts
- Refactored API route to use shared utilities
- Updated client component to use shared types and Next.js Image
- Added comprehensive package distribution structure

## Performance Improvements
- Zero JavaScript bundle impact for comment fetching logic
- Server-side data fetching with built-in Next.js caching (5min revalidation)
- Comments render immediately without loading states
- Better SEO with comments included in initial HTML

## Package Distribution Ready
- Hybrid export system supporting both client and server components
- Complete TypeScript definitions and documentation
- NPM package structure with proper peer dependencies
- Backward compatibility maintained

## Files Added
- lib/external-comments.ts: Shared data fetching utilities
- components/ExternalCommentsServer.tsx: React Server Component version
- components/external-comments/: Package distribution structure with docs

## Files Modified
- app/api/external-comments/route.ts: Refactored to use shared utilities
- components/ExternalComments.tsx: Updated to use shared types and Next.js Image
